Limit video on video page overflowing
When zoomed in at 200% on HD screens the video would overflow in the y
direction, trying to keep a 16:9 ratio. Giving the main container a
max-height so it only ever fills the remaining vertical space and then
giving the player divs of height 100% seems to fix this. I wonder if
there's a way to solve the main container only taking up the remaining
visible vertical space with the grid, rather than setting a max height

[#181328667]